yep, influxdb hates uint64's

+++ server.go
@@ -984,7 +984,7 @@ func (s *server) AddBytesSent(bytesSent uint64) {
 				InfluxDB.Point{
 					Measurement: "bandwidth",
 					Fields: map[string]interface{}{
-						"bytes_sent": s.bytesSent,
+						"bytes_sent": int64(s.bytesSent),
 					},
 				},
 			},
@@ -1006,7 +1006,7 @@ func (s *server) AddBytesReceived(bytesReceived uint64) {
 				InfluxDB.Point{
 					Measurement: "bandwidth",
 					Fields: map[string]interface{}{
-						"bytes_recv": s.bytesReceived,
+						"bytes_recv": int64(s.bytesReceived),
 					},
 				},
 			},
